Make loadJS use the same error handling as runJS
