+++ .changeset/early-rules-knock.md
@@ -1,0 +1,5 @@
+---
+'pleasantest': minor
+---
+
+Make loadJS share error mapping logic with runJS
