+++ tests/utils/external-throwing.ts
@@ -1,0 +1,3 @@
+let foo: string;
+
+throw new Error('asdf');
