+++ tests/utils/loadJS.test.ts
@@ -1,3 +1,16 @@
+import { withBrowser } from 'pleasantest';
+
 test.todo('loads from .ts file with transpiling');
-test.todo('if the file throws an error the error is source mapped');
-test.todo('if the file has a syntax error the location is source mapped');
+test(
+  'if the file throws an error the error is source mapped',
+  withBrowser(async ({ utils }) => {
+    await utils.loadJS('./external-throwing.ts');
+  }),
+);
+
+test(
+  'if the file has a syntax error the location is source mapped',
+  withBrowser(async ({ utils }) => {
+    await utils.loadJS('./external-with-syntax-error.ts');
+  }),
+);
